Move test into separate file
This test was failing on Travis when run in parallel with other tests in
the TestIntegrationCluster test. Removing the line "parallelize_me!"
from the top of the file worked, but I chose to move the test into its own
file to keep the existing tests running in parallel.
